\name{nptest-internal}

\alias{print.mcse}
\alias{print.np.cor.test}
\alias{print.np.loc.test}
\alias{print.np.reg.test}

\alias{rand.test.cor}
\alias{rand.test.lm1}
\alias{rand.test.lm2}
\alias{rand.test.one}
\alias{rand.test.two}

\alias{Tstat.cor}
\alias{Tstat.lm1}
\alias{Tstat.lm2}
\alias{Tstat.one}
\alias{Tstat.two}

\alias{Tstat.lm1.mv}
\alias{Tstat.lm2.mv}
\alias{Tstat.one.mv}
\alias{Tstat.two.mv}

\alias{Tperm.cor}
\alias{Tperm.lm1}
\alias{Tperm.lm2}
\alias{Tperm.one}
\alias{Tperm.two}

\alias{Tperm.lm1.mv}
\alias{Tperm.lm2.mv}
\alias{Tperm.one.mv}
\alias{Tperm.two.mv}

\title{
Internal 'nptest' Functions
}
\description{
Internal functions for 'nptest' package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }
