\name{StartupMessage}
\alias{StartupMessage}
\alias{nptestStartupMessage}

\title{Startup Message for nptest}

\description{
Prints the startup message when nptest is loaded. Not intended to be called by the user.
}

\details{
The `nptest' ascii start-up message was created using the taag software.
}

\references{
https://patorjk.com/software/taag/
}